backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
google-genai>=0.8.0
